requirements.txt
beautifulsoup4
pandas
tqdm
cohere
faiss-gpu
jupyterlab
streamlit
